pyproject.toml
[project]
name = "RDI-AgentBeats-TestBehaveAlign"
version = "0.0.0"
description = ""
requires-python = ">=3.13"
readme = "README.md"
license = { file = "LICENSE.md" }
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.41.0",
    "openai>=2.21.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    # Green agent runtime: test execution and mutation testing
    "pytest>=9.0.2",
    "pytest-bdd>=8.1.0",
    "mutmut>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/purple", "src/green", "src/TheBulletproofProtocol"]

[dependency-groups]
dev = [
    "complexipy>=5.2.0",
    "pymarkdownlnt>=0.9.35",
    # pytest is in main dependencies (runtime requirement for green agent)
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.2",
    "pyright>=1.1.408",
    "hypothesis>=6.151.9",
    "inline-snapshot>=0.32.2",
]
scripts = [
    "evalplus>=0.3.1",  # Only for scripts/download_evalplus.py
]

[tool.uv]
package = true
exclude-newer = "2026-02-23T00:00:00Z"

[tool.pyright]
pythonVersion = "3.13"
include = ["src", "tests"]
typeCheckingMode = "strict"
extraPaths = ["./venv/lib/python3.13/site-packages"]
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"

[tool.ruff]
target-version = "py313"
src = ["src", "tests"]
line-length = 100
exclude = [".claude", "_archive"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
# ignore = ["E203"]  # Whitespace before ':'
unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]

[tool.complexipy]
paths = ["src"]
max-complexity-allowed = 15
exclude = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--strict-markers"  # addopts = "-v --tb=short"
asyncio_mode = "auto"
# "function", "class", "module", "package", "session"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmark tests",
    "serial: marks tests to run serially (not in parallel)",
]

[tool.coverage]

[tool.coverage.run]
include = [
    "src/**/*.py",
]
# omit = []
# branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    # 'pragma: no cover',
    'raise AssertionError',
    'raise NotImplementedError',
]
omit = [
    'env/*',
    'venv/*',
    '.venv/*',
    '*/virtualenv/*',
    '*/virtualenvs/*',
    '*/tests/*',
]

# Pymarkdown configuration for markdown linting
# Rules: https://github.com/jackdewinter/pymarkdown/blob/master/docs/rules.md
[tool.pymarkdown]
log.level = "CRITICAL"
# MD013: Line too long
plugins.md013.enabled = false
# MD025: Multiple top-level headings in the same document
plugins.md025.enabled = false
# MD032: Lists should be surrounded by blank lines
# Note: Autofix pending (scheduled after v1.0.0 release
# violations must be fixed manually before that
# https://pymarkdown.readthedocs.io/en/stable/plugins/rule_md032/
plugins.md032.enabled = true
# frontmatter extension suppresses violations as of v0.9.34
# the specific rules have toe be disabled instead
# extensions.front-matter.enabled = true
# MD041: First line should be a top-level heading
plugins.md041.enabled = false
# MD022: Headings need blank lines around
# plugins.md022.enabled = false
# MD003: Inconsistent heading styles (setext vs atx)
plugins.md003.enabled = false

[tool.bumpversion]
current_version = "0.0.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
commit = true
tag = true
allow_dirty = false
ignore_missing_version = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/green/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/purple/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "README.md"
search = "version-{current_version}-58f4c2"
replace = "version-{new_version}-58f4c2"

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = """
## [Unreleased]
"""
replace = """
## [Unreleased]

## [{new_version}] - {now:%Y-%m-%d}
"""
